+++ zkvm/src/blockchain/mod.rs
@@ -1,4 +1,3 @@
 mod block;
 mod errors;
 mod state;
-pub mod utreexo;
