+++ zkvm/src/blockchain/utreexo.rs
@@ -14,1286 +14,7 @@ use core::mem;
 use merlin::Transcript;
 use std::collections::HashMap;
 
-/// Merkle hash of a node
-pub type Hash = [u8; 32];
 
-/// Absolute position of an item in the tree.
-pub type Position = u64;
 
-/// Merkle proof of inclusion of a node in a `Forest`.
-/// The exact tree is determined by the `position`, an absolute position of the item
-/// within the set of all items in the forest.
-/// Neighbors are counted from lowest to the highest.
-/// Left/right position of the neighbor is determined by the appropriate bit in `position`.
-/// (Lowest bit=1 means the first neighbor is to the left of the node.)
-/// `generation` points to the generation of the Forest to which the proof applies.
-/// `path` is None if this proof is for a newly added item that has no merkle path yet.
-#[derive(Clone, Debug)]
-pub struct Proof {
-    /// Generation of the forest to which the proof applies.
-    pub generation: u64,
 
-    /// Merkle path to the item. If missing, the proof applies to a yet-to-be-normalized forest.
-    pub path: Option<Path>,
-}
 
-/// Merkle path to the item.
-#[derive(Clone, Debug)]
-pub struct Path {
-    position: Position,
-    neighbors: Vec<Hash>,
-}
-
-/// Forest contains some number of perfect merkle binary trees
-/// and a list of newly added items.
-#[derive(Clone)]
-pub struct Forest<M: MerkleItem> {
-    generation: u64,
-    roots: [Option<NodeIndex>; 64], // roots of the trees for levels 0 to 63
-    insertions: Vec<Hash>, // new items (TBD: maybe use a fancy order-preserving HashSet later)
-    deletions: usize,
-    heap: Heap,
-    hasher: NodeHasher<M>,
-    phantom: PhantomData<M>,
-}
-
-/// Structure that helps auto-updating the proofs created for a previous generation of a forest.
-#[derive(Clone)]
-pub struct Catchup<M: MerkleItem> {
-    forest: Forest<M>,            // forest that stores the nodes
-    map: HashMap<Hash, Position>, // node hash -> new position offset for this node
-    phantom: PhantomData<M>,
-}
-
-/// Metrics of the Forest.
-#[derive(Copy, Clone, PartialEq, Debug)]
-pub struct Metrics {
-    /// Generation of the forest.
-    pub generation: u64,
-    /// Sum of capacities of all roots.
-    pub capacity: usize,
-    /// Number of deletions.
-    pub deletions: usize,
-    /// Number of insertions.
-    pub insertions: usize,
-    /// Lower-bound for amount of memory occupied in bytes.
-    pub memory: usize,
-}
-
-/// Represents an error in proof creation, verification, or parsing.
-#[derive(Fail, Clone, Debug, Eq, PartialEq)]
-pub enum UtreexoError {
-    /// This error occurs when we receive a proof that's outdated and cannot be auto-updated.
-    #[fail(display = "Item proof is outdated and must be re-created against the new state")]
-    OutdatedProof,
-
-    /// This error occurs when the merkle proof is too short or too long, or does not lead to a node
-    /// to which it should.
-    #[fail(display = "Merkle proof is invalid")]
-    InvalidProof,
-}
-
-impl<M: MerkleItem> Forest<M> {
-    /// Creates a new empty Forest.
-    pub fn new() -> Forest<M> {
-        Forest {
-            generation: 0,
-            roots: [None; 64],
-            insertions: Vec::new(),
-            deletions: 0,
-            heap: Heap::with_capacity(0),
-            hasher: NodeHasher::new(),
-            phantom: PhantomData,
-        }
-    }
-
-    /// Returns metrics data for this Forest
-    pub fn metrics(&self) -> Metrics {
-        Metrics {
-            generation: self.generation,
-            capacity: self.capacity() as usize,
-            deletions: self.deletions,
-            insertions: self.insertions.len(),
-            memory: mem::size_of::<Self>()
-                + mem::size_of::<Hash>() * self.insertions.len()
-                + mem::size_of::<Heap>()
-                + self.heap.storage.len() * mem::size_of::<PackedNode>(),
-        }
-    }
-
-    /// Verifies the item's proof of inclusion.
-    pub fn verify(&self, item: &M, proof: &Proof) -> Result<(), UtreexoError> {
-        if proof.generation != self.generation {
-            return Err(UtreexoError::OutdatedProof);
-        }
-
-        // 0. Fast check: if the proof relates to a newly added item.
-        let path = match &proof.path {
-            Some(path) => path,
-            None => {
-                let hash = self.hasher.leaf(item);
-                return self
-                    .find_insertion(&hash)
-                    .map(|_| ())
-                    .ok_or(UtreexoError::InvalidProof);
-            }
-        };
-
-        // 1. Locate the most nested node that we have under which the item.position is supposedly located.
-        let top = self.root_containing_position(path.position)?;
-
-        // 2. The proof should be of exact size from a leaf up to a tree root.
-        if path.neighbors.len() != top.level {
-            return Err(UtreexoError::InvalidProof);
-        }
-
-        // 3. Find the lowest-available node in a tree, up to which
-        //    we have to fill in the missing nodes based on the merkle proof.
-        //    And also check the higher-level neighbors in the proof.
-        let existing = self.existing_node_for_path(top, &path)?;
-
-        // If the existing node is the leaf, and it's marked as deleted - reject the proof
-        if existing.level == 0 && existing.modified {
-            return Err(UtreexoError::InvalidProof);
-        }
-
-        // 4. Now, walk the merkle proof starting with the leaf,
-        //    creating the missing nodes until we hit the bottom node.
-        let current_hash = self.hasher.leaf(item);
-        let current_hash = path
-            .walk_up(current_hash, &self.hasher)
-            .take(existing.level)
-            .fold(current_hash, |_, (parent, _children)| parent);
-
-        // 5. Check if we arrived at a correct lowest-available node.
-        if current_hash != existing.hash {
-            // We haven't met the node we expected to meet, so the proof is invalid.
-            return Err(UtreexoError::InvalidProof);
-        }
-
-        Ok(())
-    }
-
-    /// Adds a new item to the tree, appending a node to the end.
-    pub fn insert(&mut self, item: &M) -> Proof {
-        self.insertions.push(self.hasher.leaf(item));
-        Proof {
-            generation: self.generation,
-            path: None,
-        }
-    }
-
-    /// Fills in the missing nodes in the tree, and marks the item as deleted.
-    /// The algorithm minimizes amount of computation by taking advantage of the already available data.
-    pub fn delete(&mut self, item: &M, proof: &Proof) -> Result<(), UtreexoError> {
-        if proof.generation != self.generation {
-            return Err(UtreexoError::OutdatedProof);
-        }
-
-        // Determine the existing node which matches the proof, then verify the rest of the proof,
-        // and mark the relevant nodes as modified.
-
-        // 0. Fast check: if the proof relates to a newly added item, simply remove it,
-        //    so that transient items do not take up space until normalization.
-        let path = match &proof.path {
-            Some(path) => path,
-            None => {
-                // The path is missing, meaning the item must exist among the recent inserions.
-                let hash = self.hasher.leaf(item);
-                let index = self
-                    .find_insertion(&hash)
-                    .ok_or(UtreexoError::InvalidProof)?;
-                self.insertions.remove(index);
-                return Ok(());
-            }
-        };
-
-        // 1. Locate the most nested node that we have under which the item.position is supposedly located.
-        let top = self.root_containing_position(path.position)?;
-
-        // 2. The proof should be of exact size from a leaf up to a tree root.
-        if path.neighbors.len() != top.level {
-            return Err(UtreexoError::InvalidProof);
-        }
-
-        // 3. Find the lowest-available node in a tree, up to which
-        //    we have to fill in the missing nodes based on the merkle proof.
-        //    And also check the higher-level neighbors in the proof.
-        let existing = self.existing_node_for_path(top, &path)?;
-
-        // If the existing node is the leaf, and it's marked as deleted - reject the proof
-        if existing.level == 0 && existing.modified {
-            return Err(UtreexoError::InvalidProof);
-        }
-
-        // 4. Now, walk the merkle proof starting with the leaf,
-        //    creating the missing nodes until we hit the bottom node.
-        let hasher = &self.hasher;
-        let new_children = self.heap.transaction(|mut heap| {
-            let item_hash = hasher.leaf(item);
-
-            let (hash, children) = path
-                .directions()
-                .zip(path.walk_up(item_hash, hasher))
-                .enumerate()
-                .take(existing.level)
-                .fold(
-                    (item_hash, None),
-                    |(_hash, children), (i, (side, (parent_hash, (left_hash, right_hash))))| {
-                        let (left_children, right_children) = side.order(children, None);
-                        let (l, r) = (
-                            heap.allocate(left_hash, i, left_children),
-                            heap.allocate(right_hash, i, right_children),
-                        );
-                        (parent_hash, Some((l.index, r.index)))
-                    },
-                );
-
-            // 5. Check if we arrived at a correct lowest-available node.
-            if hash != existing.hash {
-                // We haven't met the node we expected to meet, so the proof is invalid.
-                return Err(UtreexoError::InvalidProof);
-            }
-
-            Ok(children)
-        })?;
-
-        // Connect children to the existing lower node.
-        let _ = self
-            .heap
-            .update(existing.index, |node| node.children = new_children);
-
-        // Update modification flag for all parents of the deleted leaf.
-        // Note: `existing` might be == `top`, so after this call top will pick up new children
-        // from the heap where they were written in the above line.
-        let top = self.heap.update(top.index, |node| node.modified = true);
-        let _ = path.iter().rev().try_fold(top, |node, (side, _)| {
-            node.children.map(|(l, r)| {
-                self.heap.update(side.choose(l, r).0, |node| {
-                    node.modified = true;
-                })
-            })
-        });
-
-        self.deletions += 1;
-
-        Ok(())
-    }
-
-    /// Normalizes the forest into minimal number of ordered perfect trees.
-    /// Returns a root of the new forst, the forest and a catchup structure.
-    pub fn normalize(self) -> (Hash, Forest<M>, Catchup<M>) {
-        // TBD: what's the best way to estimate the vector capacity from self.heap.len()?
-        let estimated_cap = self.heap.len() / 2 + self.insertions.len();
-
-        // Collect all nodes that were not modified.
-        // We delay allocation of the nodes so we don't have to mutably borrow `new_heap`
-        // in two iterators, and instead yield pairs `(hash, level)`.
-        let non_modified_nodes = self
-            .heap
-            .traverse(self.roots_iter(), |n| n.modified)
-            // 1) add pre-existing unmodified nodes...
-            .filter_map(|(_offset, node)| {
-                if !node.modified {
-                    Some((node.hash, node.level))
-                } else {
-                    None
-                }
-            })
-            // 2) ...and newly inserted nodes.
-            .chain(self.insertions.into_iter().map(|hash| (hash, 0)));
-
-        // we just consumed `self.insertions`, so let's also move out the hasher.
-        let hasher = self.hasher;
-
-        // Compute perfect roots for the new tree,
-        // joining together same-level nodes into higher-level nodes.
-        let (new_heap, new_roots) = non_modified_nodes.fold(
-            (
-                Heap::with_capacity(estimated_cap),
-                [None as Option<NodeIndex>; 64],
-            ),
-            |(mut new_heap, mut roots), (hash, level)| {
-                let mut node = new_heap.allocate(hash, level, None);
-                // If we have a left node at the same level already,
-                // merge it with the current node.
-                // Do the same with the new parent, until it lands on a unoccupied slot.
-                while let Some(i) = roots[node.level] {
-                    let left = new_heap.node_at(i);
-                    node = new_heap.allocate(
-                        hasher.intermediate(&left.hash, &node.hash),
-                        left.level + 1,
-                        Some((left.index, node.index)),
-                    );
-                    roots[left.level] = None;
-                }
-                // Place the node in the unoccupied slot.
-                roots[node.level] = Some(node.index);
-                (new_heap, roots)
-            },
-        );
-
-        let new_forest = Forest {
-            generation: self.generation + 1,
-            roots: new_roots,
-            insertions: Vec::new(), // will remain empty
-            deletions: 0,
-            heap: new_heap,
-            hasher,
-            phantom: self.phantom,
-        };
-
-        // Create a new, trimmed forest.
-        let trimmed_forest = new_forest.trim();
-        let catchup = new_forest.into_catchup();
-        let top_root = trimmed_forest.compute_root();
-
-        (top_root, trimmed_forest, catchup)
-    }
-}
-
-impl<M: MerkleItem> Catchup<M> {
-    /// Updates the proof if it's slightly out of date
-    /// (made against the previous generation of the Utreexo).
-    pub fn update_proof(&self, item: &M, proof: Proof) -> Result<Proof, UtreexoError> {
-        // If the proof is already up to date - return w/o changes
-        if proof.generation == self.forest.generation {
-            return Ok(proof);
-        }
-
-        // If the proof is not from the previous generation - fail.
-        if self.forest.generation == 0 || proof.generation != (self.forest.generation - 1) {
-            return Err(UtreexoError::OutdatedProof);
-        }
-
-        // For the newly added items `position` is irrelevant, so we create a dummy placeholder.
-        let mut path = proof.path.unwrap_or(Path {
-            position: 0,
-            neighbors: Vec::new(),
-        });
-
-        // Climb up the merkle path until we find an existing node or nothing.
-        let hash = self.forest.hasher.leaf(item);
-        let (midlevel, catchup_result) = path.walk_up(hash, &self.forest.hasher).fold(
-            (0, self.map.get(&hash)),
-            |(level, catchup_result), (p, _)| {
-                match catchup_result {
-                    Some(r) => (level, Some(r)),           // keep the found result
-                    None => (level + 1, self.map.get(&p)), // try a higher parent
-                }
-            },
-        );
-
-        // Fail early if we did not find any catchup point.
-        let position_offset = catchup_result.ok_or(UtreexoError::InvalidProof)?;
-
-        // Adjust the absolute position:
-        // keep the lowest (level+1) bits and add the stored position offset for the stored subtree
-        let mask: Position = (1 << midlevel) - 1; // L=0 -> ...00, L=1 -> ...01, L=2 -> ...11
-        path.position = position_offset + (path.position & mask);
-
-        // Remove all outdated neighbors
-        path.neighbors.truncate(midlevel);
-
-        // Find the root to which the updated position belongs
-        let root = self.forest.root_containing_position(path.position)?;
-
-        let directions = Directions {
-            position: path.position,
-            depth: root.level,
-        };
-        path.neighbors = self.forest.heap.walk_down(root, directions.rev()).fold(
-            path.neighbors,
-            |mut list, (_node, new_neighbor)| {
-                // TODO: this is not the fastest way to insert missing neighbors
-                list.insert(midlevel, new_neighbor.hash);
-                list
-            },
-        );
-
-        Ok(Proof {
-            generation: self.forest.generation,
-            path: Some(path),
-        })
-    }
-
-    /// Returns metrics data for this Catchup structure
-    pub fn metrics(&self) -> Metrics {
-        let mut metrics = self.forest.metrics();
-        metrics.memory += mem::size_of::<HashMap<Hash, Position>>()
-            + self.map.len() * (mem::size_of::<Hash>() + mem::size_of::<Position>());
-        metrics
-    }
-}
-
-// Internals
-
-impl<M: MerkleItem> Forest<M> {
-    /// Returns the index of the tree containing an item at a given position,
-    /// and the offset of that tree within the set of all items.
-    /// `position-offset` would be the position within that tree.
-    fn root_containing_position(&self, position: Position) -> Result<Node, UtreexoError> {
-        let mut offset: Position = 0;
-        for node in self.roots_iter() {
-            offset += node.capacity();
-            if position < offset {
-                return Ok(node);
-            }
-        }
-        Err(UtreexoError::InvalidProof)
-    }
-
-    /// Returns the lowest-available node for a given path and verifies the higher-level
-    /// neighbors in the path.
-    fn existing_node_for_path(&self, root: Node, path: &Path) -> Result<Node, UtreexoError> {
-        self.heap
-            .walk_down(root, path.directions().rev())
-            .zip(path.neighbors.iter().rev())
-            .try_fold(
-                root,
-                |_parent, ((node, actual_neighbor), proof_neighbor)| {
-                    if proof_neighbor != &actual_neighbor.hash {
-                        Err(UtreexoError::InvalidProof)
-                    } else {
-                        Ok(node)
-                    }
-                },
-            )
-    }
-
-    /// Capacity of the entire forest as defined by the top-level roots, excluding deletions and insertions.
-    fn capacity(&self) -> u64 {
-        self.roots_iter().map(|r| r.capacity()).sum()
-    }
-
-    /// Returns an iterator over roots of the forest,
-    /// from the highest to the lowest level.
-    fn roots_iter<'a>(&'a self) -> impl DoubleEndedIterator<Item = Node> + 'a {
-        self.roots
-            .iter()
-            .rev()
-            .filter_map(move |optional_index| optional_index.map(|index| self.heap.node_at(index)))
-    }
-
-    /// Finds an index of the hash in the list of freshly inserted items.
-    fn find_insertion(&self, hash: &Hash) -> Option<usize> {
-        self.insertions
-            .iter()
-            .enumerate()
-            .find(|&(_i, ref h)| h == &hash)
-            .map(|(i, _h)| i)
-    }
-
-    /// Trims the forest leaving only the root nodes.
-    /// Assumes the forest is normalized.
-    fn trim(&self) -> Forest<M> {
-        let mut trimmed_forest = Forest {
-            generation: self.generation,
-            roots: [None; 64],      // filled in below
-            insertions: Vec::new(), // will remain empty
-            deletions: 0,
-            heap: Heap::with_capacity(64), // filled in below
-            hasher: self.hasher.clone(),
-            phantom: self.phantom,
-        };
-        // copy the roots from the new forest to the trimmed forest
-        for root in self.roots_iter() {
-            let trimmed_root = trimmed_forest.heap.allocate(root.hash, root.level, None);
-            trimmed_forest.roots[trimmed_root.level] = Some(trimmed_root.index);
-        }
-        trimmed_forest
-    }
-
-    /// Wraps the forest into a Catchup structure
-    fn into_catchup(self) -> Catchup<M> {
-        // Traverse the tree to collect the catchup entries
-        let catchup_map = self.heap.traverse(self.roots_iter(), |_| true).fold(
-            HashMap::<Hash, Position>::new(),
-            |mut map, (offset, node)| {
-                if node.children == None {
-                    map.insert(node.hash, offset);
-                }
-                map
-            },
-        );
-        Catchup {
-            forest: self,
-            map: catchup_map,
-            phantom: PhantomData,
-        }
-    }
-
-    /// Hashes the top root of the entire forest, assuming it's normalized.
-    /// For that reason, DO NOT expose this method through the API.
-    /// Since each root is balanced, the top root is composed of n-1 pairs:
-    /// `hash(R3, hash(R2, hash(R1, R0)))`
-    fn compute_root(&self) -> Hash {
-        self.roots_iter()
-            .rev()
-            .fold(None, |optional_hash, node| {
-                if let Some(h) = optional_hash {
-                    // previous hash is of lower level, so it goes to the right
-                    Some(self.hasher.intermediate(&node.hash, &h))
-                } else {
-                    // this is the first iteration - use node's hash as-is
-                    Some(node.hash)
-                }
-            })
-            .unwrap_or(self.hasher.empty())
-    }
-}
-
-/// Index of a `Node` within a forest's heap storage.
-type NodeIndex = usize;
-
-struct NodeHasher<M: MerkleItem> {
-    t: Transcript,
-    phantom: PhantomData<M>,
-}
-
-impl<M: MerkleItem> Clone for NodeHasher<M> {
-    fn clone(&self) -> Self {
-        Self {
-            t: self.t.clone(),
-            phantom: self.phantom,
-        }
-    }
-}
-
-/// Node represents a leaf or an intermediate node in one of the trees.
-/// Leaves are indicated by `level=0`.
-/// Leaves and trimmed nodes have `children=None`.
-/// Root nodes have `parent=None`.
-#[derive(Copy, Clone, PartialEq, Debug)]
-struct Node {
-    hash: Hash,
-    index: NodeIndex,
-    level: usize,
-    modified: bool,
-    children: Option<(NodeIndex, NodeIndex)>,
-}
-
-/// Packed node as stored in memory.
-/// 32 bytes for hash, plus 13 bytes for metadata and parent and children indexes.
-/// Flags are: 6 bits for the level 0..63, 1 bit for "modified" and 1 bit for "has children".
-#[derive(Copy, Clone, PartialEq, Debug)]
-#[repr(packed)]
-struct PackedNode {
-    hash: Hash,
-    flags: u8,
-    children: (u32, u32),
-}
-
-/// Simialr to Path, but does not contain neighbors - only left/right directions
-/// as indicated by the bits in the `position`.
-#[derive(Copy, Clone, PartialEq, Debug)]
-struct Directions {
-    position: Position,
-    depth: usize,
-}
-
-#[derive(Copy, Clone, PartialEq, Debug)]
-enum Side {
-    Left,
-    Right,
-}
-
-impl Side {
-    /// Orders (current, neighbor) pair of nodes as (left, right) per `current`'s side.
-    fn order<T>(self, node: T, neighbor: T) -> (T, T) {
-        match self {
-            Side::Left => (node, neighbor),
-            Side::Right => (neighbor, node),
-        }
-    }
-
-    /// Returns (current, neighbor) pair, reversing effects of `order`.
-    fn choose<T>(self, left: T, right: T) -> (T, T) {
-        match self {
-            Side::Left => (left, right),
-            Side::Right => (right, left),
-        }
-    }
-
-    fn from_bit(bit: u8) -> Self {
-        match bit {
-            0 => Side::Left,
-            _ => Side::Right,
-        }
-    }
-}
-
-impl Path {
-    fn iter(&self) -> impl DoubleEndedIterator<Item = (Side, &Hash)> + ExactSizeIterator {
-        self.directions().zip(self.neighbors.iter())
-    }
-    fn directions(&self) -> Directions {
-        Directions {
-            position: self.position,
-            depth: self.neighbors.len(),
-        }
-    }
-    /// Returns an iterator that walks up the path
-    /// and yields parent hash and children hashes at each step.
-    fn walk_up<'a, 'b: 'a, M: MerkleItem>(
-        &'a self,
-        item_hash: Hash,
-        hasher: &'b NodeHasher<M>,
-    ) -> impl Iterator<Item = (Hash, (Hash, Hash))> + 'a {
-        self.iter()
-            .scan(item_hash, move |item_hash, (side, neighbor)| {
-                let (l, r) = side.order(*item_hash, *neighbor);
-                let p = hasher.intermediate(&l, &r);
-                *item_hash = p;
-                Some((p, (l, r)))
-            })
-    }
-}
-
-impl ExactSizeIterator for Directions {
-    fn len(&self) -> usize {
-        self.depth
-    }
-}
-
-impl Iterator for Directions {
-    type Item = Side;
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.depth == 0 {
-            return None;
-        }
-        let side = Side::from_bit((self.position & 1) as u8);
-        // kick out the lowest bit and shrink the depth
-        self.position >>= 1;
-        self.depth -= 1;
-        Some(side)
-    }
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (self.len(), Some(self.len()))
-    }
-}
-
-impl DoubleEndedIterator for Directions {
-    fn next_back(&mut self) -> Option<Self::Item> {
-        if self.depth == 0 {
-            return None;
-        }
-        self.depth -= 1;
-        // Note: we do not mask out the bit in `position` because we don't expose it.
-        // The bit is ignored implicitly by having the depth decremented.
-        let side = Side::from_bit(((self.position >> self.depth) & 1) as u8);
-        Some(side)
-    }
-}
-
-/// Iterator implementing traversal of the binary tree.
-struct TreeTraversal<'h, F>
-where
-    F: for<'n> Fn(&'n Node) -> bool,
-{
-    /// reference to the heap of nodes
-    heap: &'h Heap,
-    /// drill-down predicate - if it returns true, iterator traverses to the children.
-    /// if it returns false, the node is yielded, but its children are ignored.
-    predicate: F,
-    /// nodes in the queue - next node to be yielded is in the end of the list
-    nodes: Vec<(Position, NodeIndex)>,
-}
-
-impl<'h, F> TreeTraversal<'h, F>
-where
-    F: for<'n> Fn(&'n Node) -> bool,
-{
-    fn new(heap: &'h Heap, roots: impl IntoIterator<Item = Node>, predicate: F) -> Self {
-        let mut roots = roots.into_iter().peekable();
-        let cap = roots.size_hint().0 + 2 * roots.peek().map(|r| r.level).unwrap_or(0);
-        let mut t = TreeTraversal {
-            heap,
-            predicate,
-            nodes: Vec::with_capacity(cap),
-        };
-        let mut offset = 0;
-        for r in roots {
-            // insert in reverse order because the frontier of the iteration
-            // will be in the end of the list.
-            t.nodes.insert(0, (offset, r.index));
-            offset += r.capacity();
-        }
-        t
-    }
-}
-
-impl<'h, F> Iterator for TreeTraversal<'h, F>
-where
-    F: for<'n> Fn(&'n Node) -> bool,
-{
-    type Item = (Position, Node);
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some((offset, ni)) = self.nodes.pop() {
-            let node = self.heap.node_at(ni);
-            if (self.predicate)(&node) {
-                if let Some((li, ri)) = node.children {
-                    self.nodes.push((offset + node.capacity() / 2, ri));
-                    self.nodes.push((offset, li));
-                }
-            }
-            Some((offset, node))
-        } else {
-            None
-        }
-    }
-}
-
-/// Storage of all the nodes with methods to access them.
-#[derive(Clone)]
-struct Heap {
-    storage: Vec<PackedNode>,
-}
-
-struct HeapAllocOnly<'a> {
-    heap: &'a mut Heap,
-}
-
-impl<'a> HeapAllocOnly<'a> {
-    fn allocate(
-        &mut self,
-        hash: Hash,
-        level: usize,
-        children: Option<(NodeIndex, NodeIndex)>,
-    ) -> Node {
-        self.heap.allocate(hash, level, children)
-    }
-}
-
-impl Heap {
-    fn with_capacity(cap: usize) -> Self {
-        Heap {
-            storage: Vec::with_capacity(cap),
-        }
-    }
-
-    fn node_at(&self, i: NodeIndex) -> Node {
-        self.storage[i].unpack(i)
-    }
-
-    fn len(&self) -> usize {
-        self.storage.len()
-    }
-
-    /// Perform allocations in a transaction block: all allocations are removed
-    /// if the block fails.
-    fn transaction<T, E>(
-        &mut self,
-        closure: impl FnOnce(HeapAllocOnly) -> Result<T, E>,
-    ) -> Result<T, E> {
-        let pre_transaction_size = self.storage.len();
-        match closure(HeapAllocOnly { heap: self }) {
-            Ok(x) => Ok(x),
-            Err(e) => {
-                // undo all newly allocated items, but keep the allocated capacity for future use
-                self.storage.truncate(pre_transaction_size);
-                Err(e)
-            }
-        }
-    }
-
-    /// Allocates a node in the heap.
-    fn allocate(
-        &mut self,
-        hash: Hash,
-        level: usize,
-        children: Option<(NodeIndex, NodeIndex)>,
-    ) -> Node {
-        // make sure our indices fit in 32 bits.
-        assert!(self.storage.len() <= 0xffffffff);
-        let node = Node {
-            hash,
-            index: self.storage.len() as NodeIndex,
-            level,
-            modified: false,
-            children,
-        };
-        self.storage.push(node.pack());
-        node
-    }
-
-    fn update(&mut self, i: NodeIndex, closure: impl FnOnce(&mut Node)) -> Node {
-        let mut node = self.node_at(i);
-        closure(&mut node);
-        self.storage[i] = node.pack();
-        node
-    }
-
-    fn traverse<'h, F>(
-        &'h self,
-        roots: impl IntoIterator<Item = Node>,
-        predicate: F,
-    ) -> TreeTraversal<'h, F>
-    where
-        F: for<'n> Fn(&'n Node) -> bool,
-    {
-        TreeTraversal::new(self, roots, predicate)
-    }
-
-    /// Returns an iterator that walks the given path (from root down)
-    /// and yields current node and its neighbor.
-    /// * Root node is NOT included.
-    /// * The last node yielded by the iterator is the node without children.
-    /// * If the root has no children, iterator yields None on the first iteration.
-    fn walk_down<'a, 'b: 'a>(
-        &'a self,
-        root: Node,
-        directions: impl IntoIterator<Item = Side> + 'b,
-    ) -> impl Iterator<Item = (Node, Node)> + '_ {
-        directions
-            .into_iter()
-            .scan(root.children, move |children, side| {
-                children.map(|(li, ri)| {
-                    let (main, neighbor) = side.choose(self.node_at(li), self.node_at(ri));
-                    *children = main.children;
-                    (main, neighbor)
-                })
-            })
-    }
-}
-
-impl<M: MerkleItem> NodeHasher<M> {
-    fn new() -> Self {
-        NodeHasher {
-            t: Transcript::new(b"ZkVM.utreexo"),
-            phantom: PhantomData,
-        }
-    }
-
-    fn leaf(&self, item: &M) -> Hash {
-        let mut t = self.t.clone();
-        item.commit(&mut t);
-        let mut hash = [0; 32];
-        t.challenge_bytes(b"merkle.leaf", &mut hash);
-        hash
-    }
-
-    fn intermediate(&self, left: &Hash, right: &Hash) -> Hash {
-        let mut t = self.t.clone();
-        t.commit_bytes(b"L", left);
-        t.commit_bytes(b"R", right);
-        let mut hash = [0; 32];
-        t.challenge_bytes(b"merkle.node", &mut hash);
-        hash
-    }
-
-    fn empty(&self) -> Hash {
-        let mut t = self.t.clone();
-        let mut hash = [0; 32];
-        t.challenge_bytes(b"merkle.empty", &mut hash);
-        hash
-    }
-}
-
-impl Node {
-    /// maximum number of items in this subtree, ignoring deletions
-    fn capacity(&self) -> u64 {
-        1 << self.level
-    }
-
-    fn pack(&self) -> PackedNode {
-        debug_assert!(self.level < 64);
-
-        let modflag = if self.modified { 64 } else { 0 };
-
-        let (chflag, (l, r)) = self
-            .children
-            .map(|(l, r)| (128, (l as u32, r as u32)))
-            .unwrap_or((0, (0xffffffff, 0xffffffff)));
-
-        PackedNode {
-            hash: self.hash,
-            flags: (self.level as u8) + modflag + chflag,
-            children: (l, r),
-        }
-    }
-}
-
-impl PackedNode {
-    fn unpack(&self, index: NodeIndex) -> Node {
-        Node {
-            hash: self.hash,
-            index,
-            level: (self.flags & 63) as usize,
-            modified: (self.flags & 64) == 64,
-            children: if self.flags & 128 == 0 {
-                None
-            } else {
-                Some((self.children.0 as NodeIndex, self.children.1 as NodeIndex))
-            },
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::merkle::*;
-
-    impl MerkleItem for u64 {
-        fn commit(&self, t: &mut Transcript) {
-            t.commit_u64(b"test_item", *self);
-        }
-    }
-
-    #[test]
-    fn empty_utreexo() {
-        let forest0 = Forest::<u64>::new();
-        let metrics0 = forest0.metrics();
-        assert_eq!(metrics0.generation, 0);
-        assert_eq!(metrics0.capacity, 0);
-        assert_eq!(metrics0.insertions, 0);
-        assert_eq!(metrics0.deletions, 0);
-
-        let (root0, forest1, _catchup1) = forest0.normalize();
-        let metrics1 = forest1.metrics();
-        assert_eq!(root0, MerkleTree::root::<u64>(b"ZkVM.utreexo", &[]));
-        assert_eq!(metrics1.generation, 1);
-        assert_eq!(metrics1.capacity, 0);
-        assert_eq!(metrics1.insertions, 0);
-        assert_eq!(metrics1.deletions, 0);
-    }
-
-    #[test]
-    fn transient_items_utreexo() {
-        let mut forest0 = Forest::new();
-
-        let proof0 = forest0.insert(&0);
-        let proof1 = forest0.insert(&1);
-
-        assert_eq!(
-            forest0.metrics(),
-            Metrics {
-                generation: 0,
-                capacity: 0,
-                insertions: 2,
-                deletions: 0,
-                memory: forest0.metrics().memory,
-            }
-        );
-
-        forest0.delete(&1, &proof1).unwrap();
-        forest0.delete(&0, &proof0).unwrap();
-
-        // double spends are not allowed
-        assert_eq!(forest0.delete(&1, &proof1), Err(UtreexoError::InvalidProof));
-        assert_eq!(forest0.delete(&0, &proof0), Err(UtreexoError::InvalidProof));
-
-        assert_eq!(
-            forest0.metrics(),
-            Metrics {
-                generation: 0,
-                capacity: 0,
-                insertions: 0,
-                deletions: 0,
-                memory: forest0.metrics().memory,
-            }
-        );
-    }
-
-    #[test]
-    fn insert_to_utreexo() {
-        let mut forest0 = Forest::new();
-
-        let proofs0 = (0..6).map(|i| forest0.insert(&i)).collect::<Vec<_>>();
-
-        let (root, mut forest1, catchup1) = forest0.normalize();
-
-        assert_eq!(
-            forest1.metrics(),
-            Metrics {
-                generation: 1,
-                capacity: 6,
-                insertions: 0,
-                deletions: 0,
-                memory: forest1.metrics().memory,
-            }
-        );
-
-        assert_eq!(
-            root,
-            MerkleTree::root::<u64>(b"ZkVM.utreexo", &(0..6).collect::<Vec<_>>())
-        );
-
-        for i in 0..6u64 {
-            let deletion = forest1.delete(&i, &proofs0[i as usize]);
-            assert_eq!(deletion, Err(UtreexoError::OutdatedProof));
-        }
-
-        // update the proofs
-        let proofs1 = proofs0
-            .into_iter()
-            .enumerate()
-            .map(|(i, p)| catchup1.update_proof(&(i as u64), p).unwrap())
-            .collect::<Vec<_>>();
-
-        // after the proofs were updated, deletions should succeed
-        for i in 0..6u64 {
-            forest1.delete(&i, &proofs1[i as usize]).unwrap();
-        }
-
-        assert_eq!(
-            forest1.metrics(),
-            Metrics {
-                generation: 1,
-                capacity: 6,
-                insertions: 0,
-                deletions: 6,
-                memory: forest1.metrics().memory,
-            }
-        );
-    }
-
-    #[test]
-    fn insert_and_delete_utreexo() {
-        let mut forest0 = Forest::new();
-        let n = 6u64;
-        let proofs0 = (0..n).map(|i| forest0.insert(&i)).collect::<Vec<_>>();
-
-        let (_, forest1, catchup1) = forest0.normalize();
-
-        assert_eq!(
-            forest1.metrics(),
-            Metrics {
-                generation: 1,
-                capacity: n as usize,
-                insertions: 0,
-                deletions: 0,
-                memory: forest1.metrics().memory,
-            }
-        );
-
-        // update the proofs
-        let proofs1 = proofs0
-            .into_iter()
-            .enumerate()
-            .map(|(i, p)| catchup1.update_proof(&(i as u64), p).unwrap())
-            .collect::<Vec<_>>();
-
-        // after the proofs were updated, deletions should succeed
-
-        {
-            /* delete 0:
-                d                                       e
-                |\                                      | \
-                a   b   c      ->        b   c      ->  b   c
-                |\  |\  |\               |\  |\         |\  |\
-                0 1 2 3 4 5          x 1 2 3 4 5        2 3 4 5 1
-            */
-            let mut forest = forest1.clone();
-            forest.verify(&0u64, &proofs1[0]).unwrap();
-            forest.delete(&0u64, &proofs1[0]).unwrap();
-
-            // double spends are not allowed
-            assert_eq!(
-                forest.delete(&0, &proofs1[0]),
-                Err(UtreexoError::InvalidProof)
-            );
-
-            let (root, _, _) = forest.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[2, 3, 4, 5, 1])
-            )
-        }
-
-        {
-            /* delete 1:
-                d                                       e
-                |\                                      | \
-                a   b   c      ->        b   c      ->  b   c
-                |\  |\  |\               |\  |\         |\  |\
-                0 1 2 3 4 5          0 x 2 3 4 5        2 3 4 5 0
-            */
-            let mut forest = forest1.clone();
-            forest.verify(&1u64, &proofs1[1]).unwrap();
-            forest.delete(&1u64, &proofs1[1]).unwrap();
-
-            // double spends are not allowed
-            assert_eq!(
-                forest.delete(&1, &proofs1[1]),
-                Err(UtreexoError::InvalidProof)
-            );
-
-            let (root, _, _) = forest.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[2, 3, 4, 5, 0])
-            )
-        }
-
-        {
-            /* delete 2:
-                d                                       e
-                |\                                      | \
-                a   b   c      ->    a       c      ->  a   c
-                |\  |\  |\           |\      |\         |\  |\
-                0 1 2 3 4 5          0 1 x 3 4 5        0 1 4 5 3
-            */
-            let mut forest = forest1.clone();
-            forest.delete(&2u64, &proofs1[2]).unwrap();
-
-            let (root, _, _) = forest.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[0, 1, 4, 5, 3])
-            )
-        }
-
-        {
-            /* delete 5:
-                d                                       e
-                |\                                      | \
-                a   b   c      ->    a   b          ->  a   b
-                |\  |\  |\           |\  |\             |\  |\
-                0 1 2 3 4 5          0 1 2 3 4 x        0 1 2 3 4
-            */
-            let mut forest = forest1.clone();
-            forest.delete(&5u64, &proofs1[5]).unwrap();
-
-            let (root, _, _) = forest.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[0, 1, 2, 3, 4])
-            )
-        }
-
-        {
-            /* delete 2,3:
-                d                                       e
-                |\                                      | \
-                a   b   c      ->    a       c      ->  a   c
-                |\  |\  |\           |\      |\         |\  |\
-                0 1 2 3 4 5          0 1 x x 4 5        0 1 4 5
-            */
-            let mut forest = forest1.clone();
-            forest.delete(&2u64, &proofs1[2]).unwrap();
-            forest.delete(&3u64, &proofs1[3]).unwrap();
-
-            let (root, _, _) = forest.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[0, 1, 4, 5])
-            );
-
-            let mut forest_b = forest1.clone(); // try deletion in another order
-            forest_b.delete(&3u64, &proofs1[3]).unwrap();
-            forest_b.delete(&2u64, &proofs1[2]).unwrap();
-
-            let (root, _, _) = forest_b.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[0, 1, 4, 5])
-            );
-        }
-
-        {
-            /* delete 0,3:
-                d                                       f
-                |\                                      | \
-                a   b   c      ->            c      ->  e   c
-                |\  |\  |\                   |\         |\  |\
-                0 1 2 3 4 5          x 1 2 x 4 5        1 2 4 5
-            */
-            let mut forest = forest1.clone();
-            forest.verify(&0u64, &proofs1[0]).unwrap();
-            forest.verify(&3u64, &proofs1[3]).unwrap();
-            forest.delete(&0u64, &proofs1[0]).unwrap();
-            forest.delete(&3u64, &proofs1[3]).unwrap();
-
-            assert_eq!(
-                forest.delete(&3, &proofs1[3]),
-                Err(UtreexoError::InvalidProof)
-            );
-            assert_eq!(
-                forest.delete(&0, &proofs1[0]),
-                Err(UtreexoError::InvalidProof)
-            );
-
-            let (root, _, _) = forest.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[1, 2, 4, 5])
-            );
-        }
-
-        {
-            /* delete 0, insert 6, 7:
-                d                                          f
-                |\                                         | \
-                a   b   c      ->        b   c       ->    b   c   h
-                |\  |\  |\               |\  |\            |\  |\  |\
-                0 1 2 3 4 5          x 1 2 3 4 5 6 7       2 3 4 5 1 6 7
-            */
-            let mut forest = forest1.clone();
-            forest.verify(&0u64, &proofs1[0]).unwrap();
-            forest.delete(&0u64, &proofs1[0]).unwrap();
-            let proof6 = forest.insert(&6u64);
-            let proof7 = forest.insert(&7u64);
-
-            let (root, mut forest2, catchup) = forest.normalize();
-
-            assert_eq!(
-                root,
-                //MerkleTree::root::<u64>(b"ZkVM.utreexo", &[1, 6, 2, 3, 4, 5, 7])
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[2, 3, 4, 5, 1, 6, 7])
-            );
-
-            let proof6 = catchup.update_proof(&6u64, proof6).unwrap();
-            let proof7 = catchup.update_proof(&7u64, proof7).unwrap();
-            let proof1 = catchup.update_proof(&1u64, proofs1[1].clone()).unwrap();
-
-            /* delete 1, 7, insert :
-                 f                                        g
-                 | \                                      | \
-                 e   b   c        ->      b   c     ->    b   c
-                 |\  |\  |\               |\  |\          |\  |\
-                 1 6 2 3 4 5 7        x 6 2 3 4 5 x       2 3 4 5 6
-            */
-
-            forest2.verify(&1u64, &proof1).unwrap();
-            forest2.verify(&7u64, &proof7).unwrap();
-            forest2.delete(&1u64, &proof1).unwrap();
-            forest2.delete(&7u64, &proof7).unwrap();
-
-            let (root, mut forest3, catchup) = forest2.normalize();
-
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[2, 3, 4, 5, 6])
-            );
-
-            let proof6 = catchup.update_proof(&6u64, proof6).unwrap();
-            forest3.delete(&6u64, &proof6).unwrap();
-            let (root, _, _) = forest3.normalize();
-            assert_eq!(
-                root,
-                MerkleTree::root::<u64>(b"ZkVM.utreexo", &[2, 3, 4, 5])
-            );
-        }
-    }
-
-    #[test]
-    fn large_utreexo() {
-        let mut forest0 = Forest::new();
-        let n = 10_000u64;
-        let proofs0 = (0..n).map(|i| forest0.insert(&i)).collect::<Vec<_>>();
-
-        let (_, forest1, catchup1) = forest0.normalize();
-
-        let _proofs1 = proofs0
-            .into_iter()
-            .enumerate()
-            .map(|(i, p)| catchup1.update_proof(&(i as u64), p).unwrap())
-            .collect::<Vec<_>>();
-
-        assert!(forest1.metrics().memory < 1600);
-
-        // TODO: perform 1000 changes, normalize again and measure the memory footprint of the Catchup struct.
-    }
-}
